+++ frontend/package.json
@@ -48,6 +48,7 @@
     "@fortawesome/fontawesome-free": "^6.1.1",
     "@nivo/core": "^0.79.0",
     "@nivo/line": "^0.79.1",
+    "@nivo/scatterplot": "^0.79.1",
     "@types/d3": "^7.1.0",
     "@types/react-virtualized": "^9.21.21",
     "algoliasearch": "^4.13.0",
