+++ frontend/src/index.css
@@ -1,3 +1,8 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+@font-face {
+  font-family: 'Roboto Flex';
+  src: url(./assets/fonts/roboto_flex.ttf);
+}
